profiler/pyproject.toml
[tool.poetry]
name = "profiler"
version = "0.1.0"
description = ""
authors = ["Paul Mck <vixtir90@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
uvicorn = "^0.15.0"
fastapi = "^0.68.2"
pandas = "^1.3.3"
hydrosdk = "^3.0.0"
requests = "~2.23.0"
aiofiles = "^0.7.0"
python-multipart = "^0.0.5"
fileupload = "^0.1.5"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
flake8 = "^3.9.2"
flake8-import-order = "^0.18.1"
flake8-black = "^0.2.3"
black = "^21.7b0"
isort = "^5.9.2"
grpcio-tools = "^1.41.0"
yoyo-migrations = "^7.3.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
